requirements/base.txt
aiohttp>=1.3.3
aioredis>=0.3.0
redis
PyYAML
uvloop
colorlog

